Renamed 'core' to 'services' as it only contained services at this point.  Added submit button and placeholder
